chnaged button orientation, moved function out of utils.ts and added styles to build options
