+++ src/client/flogo/app/app-detail/app-detail.component.less
@@ -352,9 +352,9 @@
   margin-right: 10px;
 }
 
-.build-buttom-toggler {
+.build-button-dropdown-caret {
   display: inline-block;
-  transform: rotate(90deg);
+  transform: rotate(270deg);
   transform-origin: center center;
   margin: 0;
   padding: 0;
@@ -363,8 +363,4 @@
   height: 16px;
   top: 1px;
   font-size: 0.9em;
-
-  &.is-open {
-    transform: rotate(270deg);
-  }
 }
