+++ src/client/flogo/shared/utils.ts
@@ -411,15 +411,3 @@ export function getProfileType(app) {
   }
   return profileType;
 }
-
-export function getFileName(header) {
-  let filename = '';
-  if (header && header.indexOf('attachment') !== -1) {
-    const FILE_NAME_REGEX = /filename[^;=\n]*=((['"]).*?\2|[^;\n]*)/;
-    const matches = FILE_NAME_REGEX.exec(header);
-    if (matches != null && matches[1]) {
-      filename = matches[1].replace(/['"]/g, '');
-    }
-  }
-  return filename;
-}
